Add files via upload
A perfect competition for data science students who have completed an online course in machine learning and are looking to expand their skill set before trying a featured competition. The goal is to predict the sales price for each house. For each ID in the test set,  the value of the SalePrice variable must be predicted.